app/requirements/local.txt
-r base.txt # includes the base.txt requirements file

django-debug-toolbar==3.1.1
django-extensions==3.1.0
sentry-sdk==0.19.5
flower==0.9.7